# course3/3.clean_architecture/5.super_service/task3.3.5.1/swagger/swagger.yml
openapi: 3.0.0
info:
  title: Swagger users - swagger "2.0"
  description: |-
    golibrary
  version: 1.0.11
servers:
  - url: https://localhost:8080/api
paths:
  /users:
    post:
      summary: create user
      operationId: createuser
      requestBody:
        description: create user
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/user'
        required: true
      responses:
        "200":
          description: Successfuly create
        "400":
          description: Bad request
        "404":
          description: NotFound
        "500":
          description: InternalServerError
    get:
      summary: get all users
      operationId: ReadAllUser
      responses:
        "200":
          description: Successfuly
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/user'
        "400":
          description: Bad request
        "404":
          description: NotFound
        "500":
          description: InternalServerError
...
components:
  schemas:
    user:
      type: object
      properties:
        id:
          type: integer
          example: 0
        username:
          type: string
          example: "test"
        books_taken:
          type: array
          items:
            $ref: '#/components/schemas/book'
    book:
      type: object
      properties:
        id:
          type: integer
          example: 0
        name:
          type: string
          example: "Lion"
        author:
          type: string
          example: "Martin"
        author_id:
          type: integer
          example: 0
        taken_by:
          type: integer
          example: 0
...